- remove session manager async creation
