+++ Cartfile.resolved
@@ -22,7 +22,7 @@ github "wireapp/wire-ios-link-preview" "28.0.0"
 github "wireapp/wire-ios-protos" "25.0.0"
 github "wireapp/wire-ios-request-strategy" "211.0.0"
 github "wireapp/wire-ios-share-engine" "196.0.0"
-github "wireapp/wire-ios-sync-engine" "350.0.2"
+github "wireapp/wire-ios-sync-engine" "feature/SQCORE-285_RemoveSessionManagerAsyncInitialization"
 github "wireapp/wire-ios-system" "31.1.0"
 github "wireapp/wire-ios-testing" "22.0.0"
 github "wireapp/wire-ios-transport" "72.0.0"
